+++ packages/omero-utils/src/omero_utils/__init__.py
@@ -13,7 +13,7 @@ Available modules:
 
 """
 
-__version__ = "0.7.0"
+__version__ = "0.7.1"
 from omero_screen.config import set_env_vars
 
 set_env_vars()